fix: Badge 교체 및 InfoSection 삭제
