--- apps/web/src/features/expense/ui/expenseBottomSheet/infoSection/index.ts
@@ -1,2 +1,0 @@
-export { InfoSection } from './InfoSection';
-export type { InfoSectionProps, InfoBadge } from './InfoSection';
